serving-container/requirements.txt
flask
google-cloud-storage
tensorflow